pyproject.toml
[tool.poetry]
name = "product-registration-api"
version = "0.1.0"
description = ""
authors = ["matheushn99 <matheus.neres2022@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.31"
python-dotenv = "^1.0.1"
pydantic = "^2.8.2"
psycopg2-binary = "^2.9.9"
email-validator = "^2.2.0"
fastapi = "^0.111.0"
pyjwt = "^2.8.0"
pytest = "^8.2.2"
coverage = "^7.5.4"
pytest-cov = "^5.0.0"
alembic = "^1.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov=./src --cov-report=html"
testpaths = [
    "tests"
]
